fix to spatial hash; adjust build config to build all projects (hopefully)
Former-commit-id: 085eff0a593135dcfe33f6058be52a9e9a246f7f

+++ Tests/CollectionsPerformanceTest.cs
@@ -683,7 +683,7 @@ namespace Tests
             return rect;
         }
 #else
-        public static Vector2 NextVector(this IUniformRandom random, int area)
+        public static Vector2 NextVector(this IUniformRandom random, float area)
         {
             return new Vector2((float)(random.NextDouble() * area - area / 2.0),
                                (float)(random.NextDouble() * area - area / 2.0));
